Revert "Split dockerflow in own module"
This reverts commit 90b8fb3f56b8770e5f1c1dac04f2d96d7d16acfa.
